Adds interactive submit
The methodology for regular submit apparently works just fine for
the interactive submit too. Extracted methodology into separate
`submit_command` function to be reused by both submit methods.

Interactive submit cannot have a timeout -- could be an issue.

Added the definition file to the `transfer_output_files` list
so that if modifications are made during the interactive job, the
changes will be brought back to the AP.